Finally created the turn based rooms, by having the spawner check for the player and spawning the enemies, then having the spawner add each spawned enemy to a list and changing the door state to "locked". Each time an enemy dies, it'll remove itself from the created list, when the list is empty (enemies all assumed dead) the doors can be opened again.
Apart from that everything else works, only stuff to work on is editing the particle effect colours & light colour of the wand/orb projectiles & making enemies change colour depending on what they have been shot with.

From there the game will enter the Beta+ stage, in which everything will be polished, however it is expected that this will be delayed as the documentation (UML, MVC, FSM etc.) need to be drawn and the overall GDD written.